% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirmbox.R
\name{confirmbox}
\alias{confirmbox}
\title{Confirm Data Entry}
\usage{
confirmbox(input, confirm_title = "Confirm Entry",
  confirm_message_a = "You entered",
  confirm_message_b = ". Is this correct?", confirm_ok = "Yes",
  confirm_cancel = "No", ...)
}
\arguments{
\item{input}{The value provided by the user from a previous prompt}

\item{confirm_title}{The title of the input box}

\item{confirm_message_a}{The part of the message that goes before the input to be verified}

\item{confirm_message_b}{The part of the message that goes after the input to be verified}

\item{confirm_ok}{User input option that will return TRUE}

\item{confirm_cancel}{User input option that will return FALSE}

\item{...}{Used to handle arguments intended for other functions, as confirmbox() is a helper function}
}
\description{
This function provides a user input box in RStudio to confirm a value entered by the user from a previous entry.
If the user is not using RStudio, or their version of RStudio is not 1.1.67 or greater, it will use menu() to ask in the console.  Because portability.
}
\examples{
x <- showPrompt(title = "Your Name", message = "Please enter your name.", default = "Jerry")
confirmbox(input = x)
}
\keyword{internal}
