% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogbox.R
\name{dialogbox}
\alias{dialogbox}
\title{Notification Dialog Box}
\usage{
dialogbox(title, message)
}
\arguments{
\item{title}{The title for the dialog box}

\item{message}{The message for the dialog box}
}
\description{
This function is a wrapper for the showDialog function in rstudioapi.
The difference is that if the user is not using RStudio, or their version of RStudio is not 1.1.67 or greater, it will use readline to print the message to the console.  Because portability.
}
\examples{
dialogbox(title = "Warning", message = "Danger Will Robinson!")
}
\keyword{boxit}
