% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passbox.R
\name{passbox}
\alias{passbox}
\title{Password Input Box}
\usage{
passbox(prompt, ...)
}
\arguments{
\item{prompt}{The message you would like displayed in the input box}

\item{...}{Additional arguments to pass to cancelbox}
}
\description{
Sometimes you need some info from the user.  You could ask in the console, but that's easy to miss.  Why not use a prompt in RStudio!
passbox() provides an input box to get a password using blinded characters.  Of course, so does rstudioapi::askForPassword().
But what makes passbox() different is that it has a built in check so that if the user hits cancel, it will confirm they really meant to do so.
If the user does cancel, it will also cancel the script so that it does not run forward and error like crazy because it is missing vital input.
If the user is not using RStudio, or their version of RStudio is not 1.1.67 or greater, it will use readline() to ask in the console.  Because portability.
}
\examples{
x <- passbox(prompt = "Please enter your social security and bank account numbers")
}
\keyword{boxit}
