% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promptbox.R
\name{promptbox}
\alias{promptbox}
\title{Freehand User Input Box}
\usage{
promptbox(title, message, default = "", confirm = TRUE, ...)
}
\arguments{
\item{title}{The title of the input box}

\item{message}{The message asking for user input}

\item{default}{The default value that is pre-populated in the input box.  Default is "".}

\item{confirm}{Set confirm = FALSE in order to disable the confirmation step.  Default is TRUE.}

\item{...}{Additional arguments to pass to confirmbox and cancelbox}
}
\description{
Sometimes you need some info from the user.  You could ask in the console, but that's easy to miss.  Why not use a prompt in RStudio!
promptbox() provides a dialog box to get freehand user input.  Of course, so does rstudioapi::showPrompt().
But what makes promptbox() special is that has built in dialog boxes to confirm the user's input.
It also has a built in check so that if the user hits cancel, it will confirm they really meant to do so.
If the user does cancel, it will also cancel the script so that it does not run forward and error like crazy because it is missing vital input.
If the user is not using RStudio, or their version of RStudio is not 1.1.67 or greater, it will use readline() to ask in the console.  Because portability.
}
\examples{
x <- promptbox(title = "Name Please", message = "What's your name?", default = "Jerry")
}
\keyword{boxit}
