% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softstop.R
\name{softstop}
\alias{softstop}
\title{Stop a Script}
\usage{
softstop(message = NULL)
}
\arguments{
\item{message}{An error message to print when the script is stopped.  Default is NULL.}
}
\description{
A helper function that temporarily sets options(error = NULL) to stop the script without calling error handling
}
\examples{
softstop(message = "User stopped script")
}
\keyword{internal}
